Update Dockerfile
This is a Dockerfile script used to build a Docker image that runs a Java application.

The first line FROM openjdk:8 specifies the base image to use, which is the official OpenJDK 8 image.

The EXPOSE 8080 line exposes port 8080, which the Java application will be listening on.

The ADD command adds the devops-integration.jar file from the target directory (presumably the output of a Java build process) to the root directory of the Docker image, renaming it to devops-integration.jar.

Finally, the ENTRYPOINT line specifies the command to run when the Docker container is started. In this case, it runs the Java Virtual Machine (JVM) with the -jar option, passing the devops-integration.jar file as an argument. This will start the Java application and make it available on port 8080, which was exposed earlier.